비밀번호 변경, 비디오와 유저 relationship 구축, 404/403 작성

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async (req, res) => {
@@ -12,7 +13,8 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
+  //console.log(video);
   if (!video) {
     return res.render("404", { pageTitle: "비디오가 없습니다." });
   }
@@ -22,15 +24,25 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   /* exists 사용하지 않고 findById 사용해야 값을 넘겨받는다. */
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render("404", { pageTitle: "비디오가 없습니다." });
   }
+  //Edit Delete Video !== 본인일 경우에만 확인
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   return res.render("edit", { pageTitle: `수정: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
   /* findById === id 값을 넘겨받는다. */
@@ -40,6 +52,10 @@ export const postEdit = async (req, res) => {
   const video = await Video.exists({ _id: id });
   if (!video) {
     return res.status(404).render("404", { pageTitle: "비디오가 없습니다." });
+  }
+  //Edit Delete Video !== 본인일 경우에만 확인
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -55,13 +71,23 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path: videoFileUrl } = req.file;
   const { title, description, hashtags } = req.body;
+
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      videoFileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     console.log(error);
@@ -74,8 +100,22 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  //const user = await User.findById(_id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "비디오가 없습니다." });
+  }
+  //Edit Delete Video !== 본인일 경우에만 확인
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   //console.log(id);
   await Video.findByIdAndDelete(id);
+  //user.videos.splice(user.videos.indexOf(id), 1);
+  //user.save();
   return res.redirect("/");
 };
